More physics fixes. x64 fixes, bug fixes, removal of -fpermissive. Fixed ragdolls, enabled optimizations 99% of files.

+++ ivp/havana/havok/hk_base/base_types.h
@@ -55,7 +55,9 @@ typedef unsigned int		hk_uint32;
 #include <stddef.h>
 typedef size_t 			hk_size_t;  // CK: unsigned long int ..
 #else
-typedef unsigned int 	hk_size_t;
+//lwss x64 fix
+typedef unsigned long int 	hk_size_t;
+//lwss end
 #endif
 
 #define HK_BREAK (*((int *)0)) = 0
