+++ ivp/havana/havok/hk_physics/constraint/fixed/fixed_constraint.cpp
@@ -33,7 +33,7 @@ class hk_Fixed_Work
 			return addr;
 		}
 #else
-		static inline void *operator new (unsigned int size, void *addr){
+		static inline void *operator new (unsigned long int size, void *addr){
 			return addr;
 		}
 		static inline void operator delete (void *, void *){ }
