+++ ivp/havana/havok/hk_physics/constraint/hinge/hinge_constraint.cpp
@@ -29,7 +29,7 @@ struct hk_Hinge_Constraint_Work
 		return addr;
 	}
 #else
-	static inline void *operator new (unsigned int size, void *addr){
+	static inline void *operator new (unsigned long int size, void *addr){
 		return addr;
 	}
 	static inline void operator delete (void *, void *){ }
