+++ ivp/ivp_compact_builder/ivp_convex_decompositor.cxx
@@ -6,7 +6,7 @@
 
 #include <ivp_convex_decompositor.hxx>
 #include <geompack.hxx>
-
+#include <cstdint> // lwss - x64 fixes
 
 
 void IVP_Concave_Polyhedron_Face::add_offset(int offset_in) {
@@ -293,7 +293,10 @@ fclose(fp);
 	    }
 	    int j;
 	    for (j=0; j<subpart_work->offsets.len(); j++) {
-		int offset = (int)subpart_work->offsets.element_at(j)-1;
+        //lwss - x64 fixes
+		//int offset = (int)subpart_work->offsets.element_at(j)-1;
+		intptr_t offset = (intptr_t)subpart_work->offsets.element_at(j)-1;
+		//lwss end
 		IVP_U_Point *point = new IVP_U_Point();
 		point->k[0] = vcl_out[(offset*3)+0];
 		point->k[1] = vcl_out[(offset*3)+1];
