+++ ivp/ivp_compact_builder/ivp_surbuild_q12.cxx
@@ -444,9 +444,12 @@ void IVP_SurfaceBuilder_Q12::convert_model(int model)
 }
 
 
-void IVP_SurfaceBuilder_Q12::convert_node(int node)
+void IVP_SurfaceBuilder_Q12::convert_node(int64_t node)
 {
-    this->nodes.add((int *)node);
+    //lwss - x64 fixes
+    //this->nodes.add((int *)node);
+    this->nodes.add((int64_t *)node);
+    //lwss end
 
     if ( dnodes[node].children[0] > 0 ) {
 	this->convert_node(dnodes[node].children[0]);
@@ -561,9 +564,12 @@ void IVP_SurfaceBuilder_Q12::nodes_to_planes()
     int i;
     
     for (i=0; i<this->nodes.len(); i++) {
-	dplane_t *bsp_plane = &dplanes[dnodes[(int)nodes.element_at(i)].planenum];
+    //lwss - x64 fixes
+	//dplane_t *bsp_plane = &dplanes[dnodes[(int)nodes.element_at(i)].planenum];
+	dplane_t *bsp_plane = &dplanes[dnodes[(intptr_t)nodes.element_at(i)].planenum];
 	if ( i == this->nodes.len()-1 ) {
-	    if ( dnodes[(int)this->nodes.element_at(i)].children[0] == -1 ) {
+	    //if ( dnodes[(int)this->nodes.element_at(i)].children[0] == -1 ) {
+	    if ( dnodes[(intptr_t)this->nodes.element_at(i)].children[0] == -1 ) {
 		create_and_insert_plane( bsp_plane->normal[0],  bsp_plane->normal[1],  bsp_plane->normal[2], bsp_plane->dist);
 	    }
 	    else {
@@ -571,7 +577,8 @@ void IVP_SurfaceBuilder_Q12::nodes_to_planes()
 	    }
 	}
 	else {
-	    if ( dnodes[(int)this->nodes.element_at(i)].children[0] == (int)this->nodes.element_at(i+1) ) {
+	    //if ( dnodes[(int)this->nodes.element_at(i)].children[0] == (int)this->nodes.element_at(i+1) ) {
+	    if ( dnodes[(intptr_t)this->nodes.element_at(i)].children[0] == (intptr_t)this->nodes.element_at(i+1) ) {
 		create_and_insert_plane( bsp_plane->normal[0],  bsp_plane->normal[1],  bsp_plane->normal[2], bsp_plane->dist);
 	    }
 	    else {
