+++ vphysics/main.cpp
@@ -12,6 +12,7 @@
 #include "tier1/tier1.h"
 #include "ivu_vhash.hxx"
 
+#include <cstdint> //lwss - x64 fixes
 
 
 #if defined(_WIN32) && !defined(_X360)
@@ -109,8 +110,8 @@ public:
 	virtual IPhysicsEnvironment *GetActiveEnvironmentByIndex( int index );
 	virtual IPhysicsObjectPairHash *CreateObjectPairHash();
 	virtual void DestroyObjectPairHash( IPhysicsObjectPairHash *pHash );
-	virtual IPhysicsCollisionSet *FindOrCreateCollisionSet( unsigned int id, int maxElementCount );
-	virtual IPhysicsCollisionSet *FindCollisionSet( unsigned int id );
+	virtual IPhysicsCollisionSet *FindOrCreateCollisionSet( uintptr_t id, int maxElementCount );
+	virtual IPhysicsCollisionSet *FindCollisionSet( uintptr_t id );
 	virtual void DestroyAllCollisionSets();
 
 private:
@@ -176,7 +177,7 @@ void CPhysicsInterface::DestroyObjectPairHash( IPhysicsObjectPairHash *pHash )
 // holds a cache of these by id.
 // NOTE: This is stuffed into vphysics.dll as a sneaky way of sharing the memory between
 // client and server in single player.  So you can't have different client/server rules.
-IPhysicsCollisionSet *CPhysicsInterface::FindOrCreateCollisionSet( unsigned int id, int maxElementCount )
+IPhysicsCollisionSet *CPhysicsInterface::FindOrCreateCollisionSet( uintptr_t id, int maxElementCount )
 {
 	if ( !m_pCollisionSetHash )
 	{
@@ -195,11 +196,14 @@ IPhysicsCollisionSet *CPhysicsInterface::FindOrCreateCollisionSet( unsigned int 
 	return &m_collisionSets[index];
 }
 
-IPhysicsCollisionSet *CPhysicsInterface::FindCollisionSet( unsigned int id )
+IPhysicsCollisionSet *CPhysicsInterface::FindCollisionSet( uintptr_t id )
 {
 	if ( m_pCollisionSetHash )
 	{
-		int index = (int)m_pCollisionSetHash->find_elem( (void *)id );
+	    //lwss - x64 fixes
+		//int index = (int)m_pCollisionSetHash->find_elem( (void *)id );
+		intptr_t index = (intptr_t)m_pCollisionSetHash->find_elem( (void*)id );
+		//lwss end
 		if ( index > 0 )
 		{
 			Assert( index <= m_collisionSets.Count() );
